pyproject.toml
[project]
name = "flickvault"
version = "2.0.0"
description = "AI-powered movie collection manager with MCP server and web UI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.12",
    "mcp>=1.0.0",
    "anthropic>=0.40.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
flickvault-web = "app.main:run"
flickvault-mcp = "mcp_server.server:main"
